setAttributes takes vec of int32_t now

+++ include/minecraft/attribute.hpp
@@ -3,7 +3,7 @@
 
 #include <vector>
 
-extern void setAttributes(const std::vector<intptr_t> &attributeIndices); 
+extern void setAttributes(const std::vector<int32_t> &attributeIndices); 
 extern void enableVBOAttributes(const uint32_t &buffer, uint32_t upTo);
 extern void disableVBOAttributes(const uint32_t &buffer, uint32_t upTo);
 extern void enableVAOAttributes(const std::vector<uint32_t> &indices);
